[SPARK-46513][PS][TESTS] Move `BasicIndexingTests` to `pyspark.pandas.tests.indexes.*`
### What changes were proposed in this pull request?
Move `BasicIndexingTests` to `pyspark.pandas.tests.indexes.*`

### Why are the changes needed?
test code clean up

### Does this PR introduce _any_ user-facing change?
no, test-only

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44499 from zhengruifeng/ps_test_index_basic.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
